refactor: Cleanup tuneIn viewmodel
